Added Blocked and Reset Event
git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/jain-slee@22769 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ resources/isup/ratype/src/main/java/org/mobicents/slee/resources/ss7/isup/ratype/CircuitActivity.java
@@ -56,14 +56,14 @@ public class CircuitActivity implements Serializable {
 	public long getTransactionKey()
 	{
 		long currValue=dpc;
-		currValue=currValue<<14 + cic;
+		currValue=(currValue<<14) + (long)cic;
 		return currValue;
 	}
 	
 	public static long generateTransactionKey(int cic,int dpc)
 	{
 		long currValue=dpc;
-		currValue=currValue<<14 + cic;
+		currValue=(currValue<<14) + (long)cic;
 		return currValue;
 	}
 }
